+++ taskflowbackend/tests/test_plugins.py
@@ -8,7 +8,7 @@ from django.test import RequestFactory
 # Projectroles dependency
 from projectroles.app_settings import AppSettingAPI
 from projectroles.models import RoleAssignment, SODAR_CONSTANTS
-from projectroles.plugins import BackendPluginPoint
+from projectroles.plugins import BackendPluginPoint, get_backend_api
 
 # Irodsbackend dependency
 from irodsbackend.api import USER_GROUP_TEMPLATE
@@ -45,6 +45,7 @@ class ModifyAPITaskflowTestBase(TaskflowViewTestBase):
         self.req_factory = RequestFactory()
         self.request = self.req_factory.get('/')
         self.request.user = self.user
+        self.timeline = get_backend_api('timeline_backend')
 
 
 class TestPerformProjectModify(ModifyAPITaskflowTestBase):
@@ -104,7 +105,10 @@ class TestPerformProjectModify(ModifyAPITaskflowTestBase):
             event_name='project_create',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_create_inherited_members(self):
         """Test project creation in iRODS with inherited members"""
@@ -453,7 +457,10 @@ class TestRevertProjectModify(ModifyAPITaskflowTestBase):
             event_name='project_create_revert',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
 
 class TestPerformRoleModify(ModifyAPITaskflowTestBase):
@@ -492,7 +499,10 @@ class TestPerformRoleModify(ModifyAPITaskflowTestBase):
             event_name='role_update',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_create_parent(self):
         """Test creating parent assignment"""
@@ -515,7 +525,10 @@ class TestPerformRoleModify(ModifyAPITaskflowTestBase):
             event_name='role_update',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_create_parent_finder(self):
         """Test creating parent finder assignment"""
@@ -705,7 +718,10 @@ class TestPerformRoleModify(ModifyAPITaskflowTestBase):
             event_name='role_update_revert',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_revert_create_parent(self):
         """Test reverting parent role creation"""
@@ -728,7 +744,10 @@ class TestPerformRoleModify(ModifyAPITaskflowTestBase):
             event_name='role_update_revert',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_revert_create_parent_override(self):
         """Test reverting parent role creation with overriding child role"""
@@ -907,7 +926,10 @@ class TestPerformRoleDelete(ModifyAPITaskflowTestBase):
             event_name='role_delete',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_delete_parent(self):
         """Test deleting parent role"""
@@ -924,7 +946,10 @@ class TestPerformRoleDelete(ModifyAPITaskflowTestBase):
             event_name='role_delete',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_delete_parent_with_local(self):
         """Test deleting parent role with local role also set"""
@@ -991,7 +1016,10 @@ class TestPerformRoleDelete(ModifyAPITaskflowTestBase):
             event_name='role_delete_revert',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_revert_parent(self):
         """Test reverting parent role deletion"""
@@ -1011,7 +1039,10 @@ class TestPerformRoleDelete(ModifyAPITaskflowTestBase):
             event_name='role_delete_revert',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_revert_parent_with_local(self):
         """Test reverting parent role deletion with local member role"""
@@ -1109,7 +1140,10 @@ class TestPerformOwnerTransfer(ModifyAPITaskflowTestBase):
             event_name='role_owner_transfer',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_transfer_category_old_owner_finder(self):
         """Test category owner transfer with finder role for old owner"""
@@ -1248,7 +1282,10 @@ class TestPerformProjectSync(ModifyAPITaskflowTestBase):
             event_name='project_sync',
         )
         self.assertEqual(tl_events.count(), 1)
-        self.assertEqual(tl_events.first().get_status().status_type, 'OK')
+        self.assertEqual(
+            tl_events.first().get_status().status_type,
+            self.timeline.TL_STATUS_OK,
+        )
 
     def test_sync_existing(self):
         """Test sync with existing identical iRODS project (should not do anything)"""