+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Changed
     - Upgrade to iRODS v4.3.3 in CI (#1815)
     - Upgrade to python-irodsclient v2.1.0 (#2007)
     - Upgrade minimum supported iRODS version to v4.3.3 (#1815, #2007)
+    - Use constants for timeline event status types (#2010)
 - **Irodsinfo**
     - Update REST API versioning (#1936)
 - **Landingzones**